requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
alembic==1.11.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
bombaysoftwares-pysupp==1.0.4
boto3==1.28.52
botocore==1.31.52
Brotli==1.1.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
cfgv==3.3.1
charset-normalizer==3.3.0
click==8.1.3
cloudwatch==1.0.5
crontab==1.0.1
cssselect2==0.7.0
Deprecated==1.2.13
distlib==0.3.6
distro==1.8.0
et-xmlfile==1.1.0
exceptiongroup==1.1.1
filelock==3.12.0
Flask==2.3.2
Flask-Cors==3.0.10
Flask-Limiter==3.3.1
Flask-Mail==0.9.1
Flask-Migrate==2.6.0
Flask-Script==2.0.5
Flask-SQLAlchemy==3.0.3
flask-swagger-ui==4.11.1
fonttools==4.46.0
freezegun==1.2.2
frozenlist==1.4.0
google-api-core==2.12.0
google-auth==2.23.3
googleapis-common-protos==1.60.0
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
hashids==1.3.1
html5lib==1.1
httpcore==1.0.1
httplib2==0.22.0
httpx==0.25.1
identify==2.5.24
idna==3.4
importlib-metadata==6.8.0
importlib-resources==5.12.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
limits==3.5.0
Mako==1.2.4
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
multidict==6.0.4
nodeenv==1.8.0
numpy==1.24.4
oauthlib==3.2.2
openai==1.2.2
openpyxl==3.1.2
ordered-set==4.1.0
packaging==23.1
pandas==2.0.3
pdfkit==1.0.0
Pillow==10.1.0
platformdirs==3.5.1
pluggy==1.0.0
pre-commit==3.3.2
protobuf==4.24.4
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.4.2
pydantic-core==2.10.1
pydyf==0.8.0
Pygments==2.15.1
PyJWT==2.7.0
pyparsing==3.1.1
pyphen==0.14.0
pytest==7.3.1
pytest-ordering==0.6
python-dateutil==2.8.2
python-magic==0.4.27
python-slugify==8.0.1
pytz==2023.3.post1
PyYAML==6.0
random-password-generator==2.2.0
redis==4.5.5
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.3.5
rq==1.14.1
rq-scheduler==0.13.1
rsa==4.9
s3transfer==0.6.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.14
stripe==6.6.0
text-unidecode==1.3
tinycss2==1.2.1
tomli==2.0.1
tqdm==4.66.1
typing-extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
virtualenv==20.23.0
weasyprint==60.1
webencodings==0.5.1
Werkzeug==2.3.4
wrapt==1.15.0
xlrd==2.0.1
yarl==1.9.2
zipp==3.16.2
zopfli==0.2.3

